final_project/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-ai
